Moved another test, and added no reference handling
